+++ variables.tf
@@ -27,24 +27,6 @@ variable "bucket_name" {
   default     = ""
 }
 
-variable "s3_bucket_logging" {
-  description = "A map of configurations where to store logs"
-  type        = map(any)
-  default     = {}
-}
-
-variable "replication_configuration" {
-  type        = any
-  description = "Provides an independent configuration resource for S3 bucket replication configuration."
-  default     = {}
-}
-
-variable "replication_role" {
-  type        = string
-  description = "Role to use for bucket replication"
-  default     = null
-}
-
 variable "trail_name" {
   type        = string
   description = "Name for the cloudtrail"
@@ -105,16 +87,16 @@ variable "is_organization_trail" {
   default     = false
 }
 
-variable "key_deletion_window_in_days" {
-  type        = number
-  description = "The waiting period, specified in number of days. After the waiting period ends, AWS KMS deletes the KMS key. If you specify a value, it must be between 7 and 30, inclusive."
-  default     = 7
-}
-
 variable "trail_bucket_versioning_enabled" {
   type        = string
   description = "Specify whether to enable versioning for the trail bucket. Valid values are \"Enabled\" and \"Disabled\"."
   default     = "Enabled"
+}
+
+variable "custom_kms_policy" {
+  type        = string
+  description = "Add additional policies for the AWS CMK Key created by this module"
+  default     = null
 }
 
 variable "use_external_kms_key_id" {